Support send-receive

+++ app/src/main/java/one/mixin/android/ui/address/AddressManagementFragment.kt
@@ -1,10 +1,13 @@
 package one.mixin.android.ui.address
 
 import android.os.Bundle
+import android.text.Editable
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.widget.PopupMenu
+import androidx.core.view.isGone
+import androidx.core.view.isVisible
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelProviders
@@ -19,7 +22,7 @@ import one.mixin.android.ui.address.adapter.ItemCallback
 import one.mixin.android.ui.common.BaseFragment
 import one.mixin.android.ui.common.MixinBottomSheetDialogFragment
 import one.mixin.android.ui.common.PinBottomSheetDialogFragment
-import one.mixin.android.ui.common.itemdecoration.SpaceItemDecoration
+import one.mixin.android.ui.conversation.TransferFragment
 import one.mixin.android.ui.wallet.PinAddrBottomSheetDialogFragment
 import one.mixin.android.ui.wallet.PinAddrBottomSheetDialogFragment.Companion.DELETE
 import one.mixin.android.ui.wallet.PinAddrBottomSheetDialogFragment.Companion.MODIFY
@@ -27,6 +30,7 @@ import one.mixin.android.ui.wallet.TransactionsFragment
 import one.mixin.android.ui.wallet.TransactionsFragment.Companion.ARGS_ASSET
 import one.mixin.android.vo.Address
 import one.mixin.android.vo.AssetItem
+import one.mixin.android.widget.SearchView
 import javax.inject.Inject
 
 class AddressManagementFragment : BaseFragment() {
@@ -52,8 +56,9 @@ class AddressManagementFragment : BaseFragment() {
     private val asset: AssetItem by lazy {
         arguments!!.getParcelable(TransactionsFragment.ARGS_ASSET) as AssetItem
     }
+    private var addresses: List<Address>? = null
 
-    private val adapter: AddressAdapter by lazy { AddressAdapter(asset, true, true) }
+    private val adapter: AddressAdapter by lazy { AddressAdapter(asset, true) }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
             inflater.inflate(R.layout.fragment_address_management, container, false)
@@ -63,45 +68,79 @@ class AddressManagementFragment : BaseFragment() {
         title_view.left_ib.setOnClickListener { activity?.onBackPressed() }
         title_view.right_animator.setOnClickListener {
             activity?.addFragment(this@AddressManagementFragment,
-                AddressAddFragment.newInstance(asset, fromManagement = true), AddressAddFragment.TAG)
+                AddressAddFragment.newInstance(asset), AddressAddFragment.TAG)
+        }
+        empty_tv.setOnClickListener {
+            activity?.addFragment(this@AddressManagementFragment,
+                AddressAddFragment.newInstance(asset), AddressAddFragment.TAG)
         }
         addressViewModel.addresses(asset.assetId).observe(this, Observer {
-            adapter.addresses = it?.toMutableList()
+            val list = it?.toMutableList()
+            if (list.isNullOrEmpty()) {
+                empty_tv.isVisible = true
+                content_ll.isGone = true
+            } else {
+                empty_tv.isVisible = false
+                content_ll.isGone = false
+            }
+            addresses = list
+            adapter.addresses = list
         })
-        addr_rv.addItemDecoration(SpaceItemDecoration())
         val addrListener = object : AddressAdapter.SimpleAddressListener() {
             override fun onAddrLongClick(view: View, addr: Address) {
                 val popMenu = PopupMenu(activity!!, view)
                 popMenu.menuInflater.inflate(R.menu.address_mamangement_item, popMenu.menu)
                 popMenu.setOnMenuItemClickListener {
-                    showBottomSheet(addr)
+                    if (it.itemId == R.id.delete) {
+                        showBottomSheet(addr)
+                    } else if (it.itemId == R.id.edit) {
+                        activity?.addFragment(this@AddressManagementFragment,
+                            AddressAddFragment.newInstance(asset, addr, MODIFY), AddressAddFragment.TAG)
+                    }
                     return@setOnMenuItemClickListener true
                 }
                 popMenu.show()
             }
 
             override fun onAddrClick(addr: Address) {
-                activity?.addFragment(this@AddressManagementFragment,
-                    AddressAddFragment.newInstance(asset, addr, MODIFY, true), AddressAddFragment.TAG)
+                TransferFragment.newInstance(asset = asset, address = addr)
+                    .showNow(requireFragmentManager(), TransferFragment.TAG)
             }
         }
         ItemTouchHelper(ItemCallback(object : ItemCallback.ItemCallbackListener {
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder) {
+            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 val deletePos = viewHolder.adapterPosition
                 val addr = adapter.addresses!![deletePos]
-                val deleteItem = adapter.removeItem(viewHolder.adapterPosition)!!
-                val bottomSheet = showBottomSheet(addr)
-                fragmentManager?.executePendingTransactions()
-                bottomSheet.dialog.setOnDismissListener {
-                    bottomSheet.dismiss()
-                    if (!deleteSuccess) {
-                        adapter.restoreItem(deleteItem, deletePos)
+                if (direction == ItemTouchHelper.START) {
+                    val deleteItem = adapter.removeItem(viewHolder.adapterPosition)!!
+                    val bottomSheet = showBottomSheet(addr)
+                    fragmentManager?.executePendingTransactions()
+                    bottomSheet.dialog.setOnDismissListener {
+                        bottomSheet.dismiss()
+                        if (!deleteSuccess) {
+                            adapter.restoreItem(deleteItem, deletePos)
+                        }
                     }
+                } else if (direction == ItemTouchHelper.END) {
+                    adapter.notifyItemChanged(deletePos)
+                    activity?.addFragment(this@AddressManagementFragment,
+                        AddressAddFragment.newInstance(asset, addr, MODIFY), AddressAddFragment.TAG)
                 }
             }
         })).apply { attachToRecyclerView(addr_rv) }
         addr_rv.adapter = adapter
         adapter.setAddrListener(addrListener)
+        search_et.listener = object : SearchView.OnSearchViewListener {
+            override fun afterTextChanged(s: Editable?) {
+                adapter.addresses = addresses?.filter {
+                    val name = if (asset.isAccountTagAsset()) it.accountName else it.label
+                    name?.contains(s.toString(), ignoreCase = true) ?: false
+                }?.toMutableList()
+            }
+
+            override fun onSearch() {
+            }
+        }
     }
 
     private fun showBottomSheet(addr: Address): MixinBottomSheetDialogFragment {
